+++ astro.config.ts
@@ -39,6 +39,12 @@ export default defineConfig({
   markdown: {
     // @ts-ignore
     rehypePlugins,
+
+    shikiConfig: {
+      // Choose from Shiki's built-in themes (or add your own)
+      // https://github.com/shikijs/shiki/blob/main/docs/themes.md
+      theme: 'dracula',
+    },
   },
   vite: {
     plugins: [
