+++ src/components/App/Nav.astro
@@ -20,6 +20,11 @@ const links: NavLink[] = [
     title: 'Topics',
   },
   {
+    key: 'articles',
+    href: '/articles',
+    title: 'Articles',
+  },
+  {
     key: 'events',
     href: '/events',
     title: 'Events',
