+++ src/content/articles/open-source-manifesto.mdx
@@ -1,0 +1,7 @@
+---
+title: 'ISAAC Open Source Manifesto'
+description: "A gigantic part of our digitalized society relies heavily on Open Source Software. We wouldn't be here if it wasn't for the passionate developers and companies sharing their knowledge and expertise. As a digital agency, it is our duty to nurture the Open Source Software community."
+datePublished: 20200907
+externalUrl: 'https://web.archive.org/web/20211023150557/https://www.isaac.nl/en/developer-blog/open-source-manifest' 
+originalPublication: 'ISAAC Developer Blog' 
+---
