+++ src/content/config/index.ts
@@ -1,0 +1,7 @@
+import { articlesCollection } from './articles';
+import { topicsCollection } from './topics';
+
+export const collections = {
+  topics: topicsCollection,
+  articles: articlesCollection,
+};
