+++ src/pages/events/index.astro
@@ -22,7 +22,7 @@ const pastEvents = getPastEvents(events as unknown as Event[]);
 ---
 
 <BaseLayout {...baseLayoutProps}>
-  <Folio title="Events" show-title={true}>
+  <Folio title="Events" showTitle={true}>
     <section>
       <p>
         Sometimes, I'll get on a stage and talk about the stuff I like. Like rocking out, but with less noise.
